:white_check_mark: [#2242] Add test for mijn afspraken plugin
task: https://taiga.maykinmedia.nl/project/open-inwoner/task/2242
